Add jwt authentication (#5)
* Add jwt authentication

Added jwt authentication.
Added reference to user to all entities.
Altered database queries to filter by user_id.

* Change expiration env variable name

+++ src/repository/job_type_repository.rs
@@ -4,10 +4,17 @@ use super::errors::{NotFoundError, RepositoryError};
 
 type Result<T> = std::result::Result<T, RepositoryError>;
 
-pub async fn get_all_async(pool: &sqlx::Pool<sqlx::Postgres>) -> Result<Vec<JobType>> {
-    let result = match sqlx::query_as!(JobType, "SELECT * FROM tb_job_types")
-        .fetch_all(pool)
-        .await
+pub async fn get_all_async(
+    pool: &sqlx::Pool<sqlx::Postgres>,
+    user_id: i32,
+) -> Result<Vec<JobType>> {
+    let result = match sqlx::query_as!(
+        JobType,
+        "SELECT * FROM tb_job_types WHERE user_id = $1",
+        user_id
+    )
+    .fetch_all(pool)
+    .await
     {
         Ok(result) => Ok(result),
         Err(e) => Err(RepositoryError::InternalError(e)),
@@ -16,10 +23,19 @@ pub async fn get_all_async(pool: &sqlx::Pool<sqlx::Postgres>) -> Result<Vec<JobT
     return result;
 }
 
-pub async fn get_by_id_async(pool: &sqlx::Pool<sqlx::Postgres>, id: i32) -> Result<JobType> {
-    let result = match sqlx::query_as!(JobType, "SELECT * FROM tb_job_types WHERE id = $1", id)
-        .fetch_one(pool)
-        .await
+pub async fn get_by_id_async(
+    pool: &sqlx::Pool<sqlx::Postgres>,
+    user_id: i32,
+    id: i32,
+) -> Result<JobType> {
+    let result = match sqlx::query_as!(
+        JobType,
+        "SELECT * FROM tb_job_types WHERE id = $1 AND user_id = $2",
+        id,
+        user_id
+    )
+    .fetch_one(pool)
+    .await
     {
         Ok(result) => Ok(result),
         Err(sqlx::Error::RowNotFound) => Err(RepositoryError::from(NotFoundError::ById(id))),
@@ -29,10 +45,19 @@ pub async fn get_by_id_async(pool: &sqlx::Pool<sqlx::Postgres>, id: i32) -> Resu
     return result;
 }
 
-pub async fn get_by_name_async(pool: &sqlx::Pool<sqlx::Postgres>, name: &str) -> Result<JobType> {
-    let result = match sqlx::query_as!(JobType, "SELECT * FROM tb_job_types WHERE name = $1", name)
-        .fetch_one(pool)
-        .await
+pub async fn get_by_name_async(
+    pool: &sqlx::Pool<sqlx::Postgres>,
+    user_id: i32,
+    name: &str,
+) -> Result<JobType> {
+    let result = match sqlx::query_as!(
+        JobType,
+        "SELECT * FROM tb_job_types WHERE name = $1 AND user_id = $2",
+        name,
+        user_id
+    )
+    .fetch_one(pool)
+    .await
     {
         Ok(result) => Ok(result),
         Err(sqlx::Error::RowNotFound) => Err(RepositoryError::from(NotFoundError::ByProperty(
@@ -48,8 +73,9 @@ pub async fn get_by_name_async(pool: &sqlx::Pool<sqlx::Postgres>, name: &str) ->
 pub async fn create_async(pool: &sqlx::Pool<sqlx::Postgres>, job_type: JobType) -> Result<JobType> {
     let result = match sqlx::query_as!(
         JobType,
-        "INSERT INTO tb_job_types (name, modified_at) VALUES ($1, $2) RETURNING *",
+        "INSERT INTO tb_job_types (name, user_id, modified_at) VALUES ($1, $2, $3) RETURNING *",
         job_type.name,
+        job_type.user_id,
         job_type.modified_at
     )
     .fetch_one(pool)
@@ -69,26 +95,33 @@ pub async fn update_async(
 ) -> Result<JobType> {
     let result = match sqlx::query_as!(
         JobType,
-        "UPDATE tb_job_types SET name = $1, modified_at = $2 WHERE id = $3 RETURNING *",
+        "UPDATE tb_job_types SET name = $1, modified_at = $2 WHERE id = $3 AND user_id = $4 RETURNING *",
         job_type.name,
         job_type.modified_at,
-        id
+        id,
+        job_type.user_id
     )
     .fetch_one(pool)
     .await
     {
         Ok(result) => Ok(result),
+        Err(sqlx::Error::RowNotFound) => Err(RepositoryError::from(NotFoundError::ById(id))),
         Err(e) => Err(RepositoryError::InternalError(e)),
     };
 
     return result;
 }
 
-pub async fn delete_async(pool: &sqlx::Pool<sqlx::Postgres>, id: i32) -> Result<JobType> {
+pub async fn delete_async(
+    pool: &sqlx::Pool<sqlx::Postgres>,
+    user_id: i32,
+    id: i32,
+) -> Result<JobType> {
     let result = match sqlx::query_as!(
         JobType,
-        "DELETE FROM tb_job_types WHERE id = $1 RETURNING *",
-        id
+        "DELETE FROM tb_job_types WHERE id = $1 AND user_id = $2 RETURNING *",
+        id,
+        user_id
     )
     .fetch_one(pool)
     .await
@@ -103,11 +136,13 @@ pub async fn delete_async(pool: &sqlx::Pool<sqlx::Postgres>, id: i32) -> Result<
 
 pub async fn check_duplicate_by_name(
     pool: &sqlx::Pool<sqlx::Postgres>,
+    user_id: i32,
     name: &str,
 ) -> Result<bool> {
     let result = match sqlx::query!(
-        "SELECT EXISTS(SELECT 1 FROM tb_job_types WHERE name = $1)",
-        name
+        "SELECT EXISTS(SELECT 1 FROM tb_job_types WHERE name = $1 AND user_id = $2)",
+        name,
+        user_id
     )
     .fetch_one(pool)
     .await
